requirements.txt
Django==4.2.7
reportlab==4.0.4
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.7
python-decouple==3.8
dj-database-url==2.1.0
